move non-models out of model module
update import
